+++ src/gameobjects/bitmaptext/dynamic/DynamicBitmapText.js
@@ -111,7 +111,7 @@ var DynamicBitmapText = new Class({
          * @type {string}
          * @since 3.0.0
          */
-        this.text = (Array.isArray(text)) ? text.join('\n') : text;
+        this.text = '';
 
         /**
          * [description]
@@ -131,6 +131,8 @@ var DynamicBitmapText = new Class({
          * @since 3.5.0
          */
         this.letterSpacing = 0;
+
+        this.setText(text);
 
         this.setTexture(entry.texture, entry.frame);
         this.setPosition(x, y);